# Conflicts:
#	app/build.gradle.kts

+++ app/src/main/java/com/project/veganlife/mypage/ui/adapter/MypagePostedFeedAdapter.kt
@@ -55,7 +55,7 @@ class MypagePostedFeedAdapter :
 
                 clMypageLayout.setOnClickListener {
                     val bundle = Bundle()
-                    bundle.putInt("feedId", postedFeed.id)
+                    bundle.putLong("feedId", postedFeed.id)
                     itemView.findNavController().navigate(
                         R.id.action_mypagePostedFeedFragment_to_communityDetailFeedFragment,
                         bundle
